refactor: code analyze.
